code/requirements.txt
pandas
matplotlib
seaborn
plotly
kaleido
Pillow
